Move device function to core to fix circular import error
